nexed/python-novice/07-level/Tabel-overzetten-82a16517e933-fff69f363145/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
mysql-connector-python==9.2.0
requests==2.32.3
schedule==1.2.2
urllib3==2.3.0
